# Conflicts:
#	mmseg/__init__.py
#	mmseg/datasets/__init__.py
#	mmseg/models/backbones/__init__.py
#	mmseg/models/backbones/vit.py
#	mmseg/models/decode_heads/__init__.py
#	mmseg/models/decode_heads/dpt_head.py
#	mmseg/models/necks/multilevel_neck.py
#	mmseg/models/utils/__init__.py
#	mmseg/models/utils/embed.py

+++ mmseg/models/losses/cross_entropy_loss.py
@@ -19,7 +19,7 @@ def cross_entropy(pred,
     # If given, has to be a Tensor of size C element-wise losses
     loss = F.cross_entropy(
         pred,
-        label,
+        label.long(),
         weight=class_weight,
         reduction='none',
         ignore_index=ignore_index)
